refactor(reth): Extract `Cli` into `ethereum-cli` crate
