--- bin/reth/src/commands/mod.rs
@@ -1,3 +1,0 @@
-//! This contains all of the `reth` commands
-
-pub mod debug_cmd;
