+++ crates/ethereum/cli/src/debug_cmd/merkle.rs
@@ -1,5 +1,4 @@
 //! Command for debugging merkle tree calculation.
-use crate::{args::NetworkArgs, providers::ExecutionOutcome, utils::get_single_header};
 use alloy_eips::BlockHashOrNumber;
 use backon::{ConstantBuilder, Retryable};
 use clap::Parser;
@@ -13,10 +12,12 @@ use reth_consensus::{Consensus, ConsensusError};
 use reth_db_api::{cursor::DbCursorRO, tables, transaction::DbTx};
 use reth_ethereum_primitives::EthPrimitives;
 use reth_evm::execute::{BlockExecutorProvider, Executor};
+use reth_execution_types::ExecutionOutcome;
 use reth_network::{BlockDownloaderProvider, NetworkHandle};
 use reth_network_api::NetworkInfo;
 use reth_network_p2p::full_block::FullBlockClient;
 use reth_node_api::{BlockTy, NodePrimitives};
+use reth_node_core::{args::NetworkArgs, utils::get_single_header};
 use reth_node_ethereum::{consensus::EthBeaconConsensus, EthExecutorProvider};
 use reth_provider::{
     providers::ProviderNodeTypes, BlockNumReader, BlockWriter, ChainSpecProvider,
